gwpy.table: improve tests of PyCBC I/O
- migrate all tests to the separate `test_io_pycbc.py` module
- add more tests
