+++ gwpy/table/tests/test_table.py
@@ -34,7 +34,6 @@ from astropy import units
 from astropy.io.ascii import InconsistentTableError
 from astropy.table import vstack
 
-from ...frequencyseries import FrequencySeries
 from ...io import ligolw as io_ligolw
 from ...segments import (Segment, SegmentList)
 from ...testing import utils
@@ -686,193 +685,6 @@ class TestEventTable(TestTable):
             )
         assert str(exc.value) == (
             'No column names found in {} header'.format(fmtname)
-        )
-
-    @pytest.fixture
-    def pycbclivetable(self):
-        return self.create(
-            100,
-            names=[
-                'a',
-                'b',
-                'c',
-                'chisq',
-                'd',
-                'e',
-                'f',
-                'mass1',
-                'mass2',
-                'snr',
-            ],
-        )
-
-    @staticmethod
-    @pytest.fixture
-    def pycbclivepsd():
-        return FrequencySeries(random.randn(1000), df=1)
-
-    @pytest.fixture
-    def pycbclivefile(self, tmp_path, pycbclivetable, pycbclivepsd):
-        # create table
-        loudest = (pycbclivetable['snr'] > 500).nonzero()[0]
-
-        # manually create pycbc_live-format HDF5 file
-        tmp = tmp_path / "X1-Live-0-0.hdf"
-        # write table in pycbc_live format (by hand)
-        with h5py.File(tmp, "w") as h5f:
-            group = h5f.create_group('X1')
-            for col in pycbclivetable.columns:
-                group.create_dataset(data=pycbclivetable[col], name=col)
-            group.create_dataset('loudest', data=loudest)
-            group.create_dataset('psd', data=pycbclivepsd.value)
-            group['psd'].attrs['delta_f'] = pycbclivepsd.df.to('Hz').value
-        return tmp
-
-    def test_read_pycbc_live(self, pycbclivetable, pycbclivefile):
-        """Check that we can read a PyCBC-Live file
-        """
-        table = self.TABLE.read(pycbclivefile, format="hdf5.pycbc_live")
-        utils.assert_table_equal(pycbclivetable, table)
-        assert table.meta['ifo'] == 'X1'
-
-    def test_read_pycbc_live_kwargs(self, pycbclivetable, pycbclivefile):
-        """Check that we can read a PyCBC-Live file using keywords
-        """
-        table = self.TABLE.read(
-            pycbclivefile,
-            format='hdf5.pycbc_live',
-            ifo='X1',
-        )
-        utils.assert_table_equal(pycbclivetable, table)
-
-    def test_read_pycbc_live_loudest(self, pycbclivetable, pycbclivefile):
-        """Check that we can read the `loudest` table from a PyCBC-Live file
-        """
-        table = self.TABLE.read(
-            pycbclivefile,
-            format="hdf5.pycbc_live",
-            loudest=True,
-        )
-        utils.assert_table_equal(pycbclivetable.filter('snr > 500'), table)
-
-    def test_read_pycbc_live_extended_metadata(
-            self,
-            pycbclivetable,
-            pycbclivepsd,
-            pycbclivefile,
-    ):
-        """Check that we can read extended metadata from a PyCBC-Live file
-        """
-        table = self.TABLE.read(
-            pycbclivefile,
-            format="hdf5.pycbc_live",
-            extended_metadata=True,  # default
-        )
-        utils.assert_table_equal(pycbclivetable, table)
-        utils.assert_array_equal(
-            table.meta['loudest'],
-            (pycbclivetable['snr'] > 500).nonzero()[0],
-        )
-        utils.assert_quantity_sub_equal(
-            table.meta['psd'],
-            pycbclivepsd,
-            exclude=['name', 'channel', 'unit', 'epoch'])
-
-    def test_read_pycbc_live_extended_metadata_false(
-            self,
-            pycbclivetable,
-            pycbclivefile,
-    ):
-        """Check that we can read a PyCBC-Live file without extended metadata
-        """
-        # check extended_metadata=False works
-        table = self.TABLE.read(
-            pycbclivefile,
-            format="hdf5.pycbc_live",
-            extended_metadata=False,
-        )
-        assert table.meta == {'ifo': 'X1'}
-
-    def test_read_pycbc_live_multiple_ifos(
-            self,
-            pycbclivetable,
-            pycbclivefile,
-    ):
-        """Check that we can handle multiple IFOs in a PyCBC-Live file
-        """
-        with h5py.File(pycbclivefile, "r+") as h5f:
-            h5f.create_group('Z1')
-        with pytest.raises(ValueError) as exc:
-            self.TABLE.read(pycbclivefile, format="hdf5.pycbc_live")
-        assert str(exc.value).startswith(
-            'PyCBC live HDF5 file contains dataset groups')
-
-        # but check that we can still read the original
-        table = self.TABLE.read(
-            pycbclivefile,
-            format='hdf5.pycbc_live',
-            ifo='X1',
-        )
-        utils.assert_table_equal(pycbclivetable, table)
-
-    def test_read_pycbc_live_processed_columns(
-            self,
-            pycbclivetable,
-            pycbclivefile,
-    ):
-        """Check that we can read processed columns from a PyCBC-Live file
-        """
-        # assert processed colums works
-        table = self.TABLE.read(
-            pycbclivefile,
-            format="hdf5.pycbc_live",
-            ifo="X1",
-            columns=["mchirp", "new_snr"],
-        )
-        mchirp = (
-            (pycbclivetable['mass1'] * pycbclivetable['mass2']) ** (3/5.)
-            / (pycbclivetable['mass1'] + pycbclivetable['mass2']) ** (1/5.)
-        )
-        utils.assert_array_equal(table['mchirp'], mchirp)
-
-    def test_read_pycbc_live_selection_columns(
-            self,
-            pycbclivetable,
-            pycbclivefile,
-    ):
-        """Check that the selection and columns kwargs work when
-        reading from a PyCBC-Live file
-        """
-        # test with selection and columns
-        table = self.TABLE.read(
-            pycbclivefile,
-            format='hdf5.pycbc_live',
-            ifo='X1',
-            selection='snr>.5',
-            columns=("a", "b", "mass1"),
-        )
-        utils.assert_table_equal(
-            table,
-            filter_table(pycbclivetable, 'snr>.5')[("a", "b", "mass1")],
-        )
-
-    def test_read_pycbc_live_regression_1081(
-            self,
-            pycbclivetable,
-            pycbclivefile,
-    ):
-        """Check against regression of gwpy/gwpy#1081
-        """
-        table = self.TABLE.read(
-            pycbclivefile,
-            format='hdf5.pycbc_live',
-            ifo='X1',
-            selection='snr>.5',
-            columns=("a", "b", "snr"),
-        )
-        utils.assert_table_equal(
-            table,
-            filter_table(pycbclivetable, 'snr>.5')[("a", "b", "snr")],
         )
 
     @pytest.fixture
